requirements.txt
pytest
mysql-connector
translate
flask
flask_cors